get an initial loop working
