+++ vanilla/src/game.js
@@ -1,10 +1,8 @@
 import { createApp } from './components/app/app.js';
 
 export function start() {
-  const app = createApp({
+  createApp({
     row: 25,
     col: 25,
   });
-
-  document.body.appendChild(app.render());
 }
